+++ grafanalib/core.py
@@ -542,6 +542,20 @@ class Annotations(object):
 
 
 @attr.s
+class DataLink(object):
+    title = attr.ib()
+    linkUrl = attr.ib(default="", validator=instance_of(str))
+    isNewTab = attr.ib(default=False, validator=instance_of(bool))
+
+    def to_json_data(self):
+        return {
+            'title': self.title,
+            'url': self.linkUrl,
+            'targetBlank': self.isNewTab,
+        }
+
+
+@attr.s
 class DataSourceInput(object):
     name = attr.ib()
     label = attr.ib()
@@ -1014,6 +1028,7 @@ class Graph(object):
     """
     Generates Graph panel json structure.
 
+    :param dataLinks: list of data links hooked to datapoints on the graph
     :param dataSource: DataSource's name
     :param minSpan: Minimum width for each panel
     :param repeat: Template's name to repeat Graph on
@@ -1023,6 +1038,7 @@ class Graph(object):
     targets = attr.ib()
     aliasColors = attr.ib(default=attr.Factory(dict))
     bars = attr.ib(default=False, validator=instance_of(bool))
+    dataLinks = attr.ib(default=attr.Factory(list))
     dataSource = attr.ib(default=None)
     description = attr.ib(default=None)
     editable = attr.ib(default=True, validator=instance_of(bool))
@@ -1083,6 +1099,9 @@ class Graph(object):
             'links': self.links,
             'minSpan': self.minSpan,
             'nullPointMode': self.nullPointMode,
+            'options': {
+                'dataLinks': self.dataLinks,
+            },
             'percentage': self.percentage,
             'pointradius': self.pointRadius,
             'points': self.points,
@@ -1692,6 +1711,7 @@ class BarGauge(object):
     :param allValue: If All values should be shown or a Calculation
     :param cacheTimeout: metric query result cache ttl
     :param calc: Calculation to perform on metrics
+    :param dataLinks: list of data links hooked to datapoints on the graph
     :param dataSource: Grafana datasource name
     :param decimals: override automatic decimal precision for legend/tooltips
     :param description: optional panel description
@@ -1729,6 +1749,7 @@ class BarGauge(object):
     allValues = attr.ib(default=False, validator=instance_of(bool))
     cacheTimeout = attr.ib(default=None)
     calc = attr.ib(default=GAUGE_CALC_MEAN)
+    dataLinks = attr.ib(default=attr.Factory(list))
     dataSource = attr.ib(default=None)
     decimals = attr.ib(default=None)
     description = attr.ib(default=None)
@@ -1801,6 +1822,7 @@ class BarGauge(object):
                         "min": self.min,
                         "title": self.label,
                         "unit": self.format,
+                        "links": self.dataLinks,
                     },
                     "limit": self.limit,
                     "mappings": self.valueMaps,
@@ -1830,6 +1852,7 @@ class GaugePanel(object):
     :param allValue: If All values should be shown or a Calculation
     :param cacheTimeout: metric query result cache ttl
     :param calc: Calculation to perform on metrics
+    :param dataLinks: list of data links hooked to datapoints on the graph
     :param dataSource: Grafana datasource name
     :param decimals: override automatic decimal precision for legend/tooltips
     :param description: optional panel description
@@ -1865,6 +1888,7 @@ class GaugePanel(object):
     allValues = attr.ib(default=False, validator=instance_of(bool))
     cacheTimeout = attr.ib(default=None)
     calc = attr.ib(default=GAUGE_CALC_MEAN)
+    dataLinks = attr.ib(default=attr.Factory(list))
     dataSource = attr.ib(default=None)
     decimals = attr.ib(default=None)
     description = attr.ib(default=None)
@@ -1922,6 +1946,7 @@ class GaugePanel(object):
                         "min": self.min,
                         "title": self.label,
                         "unit": self.format,
+                        "links": self.dataLinks,
                     },
                     "limit": self.limit,
                     "mappings": self.valueMaps,